requirements.txt
matplotlib==3.7.1
munch==2.5.0
numpy==1.24.2
Pillow==9.4.0
plotly==5.9.0
scikit_learn==1.2.2
scipy==1.10.0
seaborn==0.12.2
setuptools==65.6.3
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
